Refactor ApiTestCase class hierarchy
Also added a new base ScenarioTest class, which extends
MethodTest;  Method and Scenario tests cases have access to gRPC
service stubs, but end to end test cases should never use them.
