+++ apitest/src/main/java/bisq/apitest/method/WalletProtectionTest.java
@@ -6,9 +6,7 @@ import lombok.extern.slf4j.Slf4j;
 
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 
 import static bisq.apitest.config.BisqAppConfig.alicedaemon;
@@ -21,9 +19,6 @@ import bisq.apitest.annotation.Order;
 @Slf4j
 @RunWith(OrderedRunner.class)
 public class WalletProtectionTest extends MethodTest {
-
-    @Rule
-    public ExpectedException exceptionRule = ExpectedException.none();
 
     @BeforeClass
     public static void setUp() {
