+++ apitest/src/main/java/bisq/apitest/scenario/ScenarioTest.java
@@ -1,0 +1,6 @@
+package bisq.apitest.scenario;
+
+import bisq.apitest.method.MethodTest;
+
+public class ScenarioTest extends MethodTest {
+}
